use bitmask approach instead

+++ src/subsystems/vision/visionio.py
@@ -3,8 +3,7 @@ from enum import Enum
 from typing import List
 from pykit.autolog import autolog
 from wpimath.geometry import Pose3d, Transform3d
-from wpiutil.wpistruct import make_wpistruct
-from wpiutil.wpistruct import uint64
+from wpiutil.wpistruct import make_wpistruct, uint32
 
 
 class ObservationType(Enum):
@@ -21,7 +20,7 @@ class VisionSubsystemPoseObservation:
     pose: Pose3d = field(default_factory=Pose3d)
     ambiguity: float = 0
     tagCount: int = 0
-    tagsList: uint64 = uint64(0)
+    tagsList: uint32 = uint32(0)
     averageTagDistance: float = 0
     observationType: int = ObservationType.PHOTONVISION.value
 
@@ -34,7 +33,7 @@ class VisionSubsystemTurretedPoseObservation:
     fieldToTurret: Transform3d = field(default_factory=Transform3d)
     ambiguity: float = 0
     tagCount: int = 0
-    tagsList: uint64 = uint64(0)
+    tagsList: uint32 = uint32(0)
     averageTagDistance: float = 0
     observationType: int = ObservationType.PHOTONVISION.value
